chore:refact code structure
