+++ src/lib.rs
@@ -1,5 +1,5 @@
-mod csvs;
+mod opts;
 mod process;
 
-pub use csvs::{CsvOpts, Opts, Player, Subcommand};
+pub use opts::{CsvOpts, Opts, Subcommand};
 pub use process::process_csv_2_json;
