+++ src/process.rs
@@ -1,6 +1,24 @@
-use crate::csvs::Player;
 use csv::Reader;
+use serde::{Deserialize, Serialize};
 use std::fs;
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct Player {
+    #[serde(rename = "Name")]
+    name: String,
+
+    #[serde(rename = "Position")]
+    position: String,
+
+    #[serde(rename = "DOB")]
+    dob: String,
+
+    #[serde(rename = "Nationality")]
+    nationality: String,
+
+    #[serde(rename = "Kit Number")]
+    kit: u8,
+}
 
 pub fn process_csv_2_json(filepath: &str, outputfilepath: &str) -> anyhow::Result<()> {
     let mut reader = Reader::from_path(filepath)?;
